+++ src/main/java/me/despical/inventoryframework/pane/Fillable.java
@@ -37,6 +37,25 @@ public interface Fillable {
     void fillHorizontallyWith(@NotNull ItemStack itemStack, int line);
 
     /**
+     * Fills specified row line vertically with given {@code itemStack}
+     *
+     * @param itemStack The {@link ItemStack} to fill the empty space with
+     * @param line      Line to fill with {@code itemStack}
+     * @param action    The action called whenever an interaction with the item happens
+     * @since 1.0.8
+     */
+    void fillVerticallyWith(@NotNull ItemStack itemStack, int line, @Nullable Consumer<InventoryClickEvent> action);
+
+    /**
+     * Fills specified row line vertically with given {@code itemStack}
+     *
+     * @param itemStack The {@link ItemStack} to fill the empty space with
+     * @param line      Line to fill with {@code itemStack}
+     * @since 1.0.8
+     */
+    void fillVerticallyWith(@NotNull ItemStack itemStack, int line);
+
+    /**
      * Fills inventory borders with given {@code itemStack}
      *
      * @param itemStack The {@link ItemStack} to fill the empty space with